feat: ssr download portal
