+++ src/domain/props/CountryReportsProps.ts
@@ -1,0 +1,5 @@
+import { CountryCodesData } from '../entities/country/CountryCodesData';
+
+export default interface CountryReportsProps {
+  countryCodesData: CountryCodesData[];
+}
