+++ src/domain/props/PdfPreviewProps.ts
@@ -1,0 +1,9 @@
+import { PdfFile } from './PdfViewerProps';
+
+export default interface PdfPreviewProps {
+  isModalOpen: boolean;
+  toggleModal: () => void;
+  pdfFile: PdfFile | null;
+  error: string | null;
+  onDownloadPdf: () => void;
+}
